client -> execution refactor: moved tests over to dedicated execution test file
